revert to prior memcached caching method

+++ lib/util/SRA_Cache.php
@@ -94,13 +94,13 @@ class SRA_Cache {
 	 * @return mixed
 	 */
   function cacheIsset($name, $modtime=FALSE) {
-    global $argc, $memcached, $memcachedKeys;
+    global $argc, $memcached;
     if (SRA_CACHE_DEBUG) SRA_Error::logError('SRA_Cache::cacheIsset - invoked for "' . $name . '"', __FILE__, __LINE__);
     
     // use memcached if global variable $memcached exists
     if (isset($memcached) && class_exists('Memcached') && get_class($memcached) == 'Memcached') {
-      if (!is_array($memcachedKeys)) $memcachedKeys = $memcached->getAllKeys();
-      return in_array($name, $memcachedKeys) || $memcached->get($name) ? TRUE : FALSE;
+      $memcached->get($name);
+      return $memcached->getResultCode() != Memcached::RES_NOTFOUND;
     }
     
     // use APC if present
@@ -132,14 +132,11 @@ class SRA_Cache {
 	 * @return boolean
 	 */
   function deleteCache($name) {
-    global $argc, $memcached, $memcachedKeys;
+    global $argc, $memcached;
     if (SRA_CACHE_DEBUG) SRA_Error::logError('SRA_Cache::deleteCache - invoked for "' . $name . '"', __FILE__, __LINE__);
     
     // use memcached if global variable $memcached exists
-    if (isset($memcached) && class_exists('Memcached') && get_class($memcached) == 'Memcached') {
-      if (is_array($memcachedKeys) && in_array($name, $memcachedKeys)) unset($memcachedKeys[array_search($name, $memcachedKeys)]);
-      return $memcached->delete($name);
-    }
+    if (isset($memcached) && class_exists('Memcached') && get_class($memcached) == 'Memcached') return $memcached->delete($name);
     
     // use APC if present
     if (!isset($argc) && function_exists('apc_delete')) return apc_delete($name);
@@ -177,9 +174,7 @@ class SRA_Cache {
     if (SRA_CACHE_DEBUG) SRA_Error::logError('SRA_Cache::getCache - invoked for "' . $name . '"', __FILE__, __LINE__);
     
     // use memcached if global variable $memcached exists
-    if (isset($memcached) && class_exists('Memcached') && get_class($memcached) == 'Memcached') {
-      return $memcached->get($name);
-    }
+    if (isset($memcached) && class_exists('Memcached') && get_class($memcached) == 'Memcached') return $memcached->get($name);
     
     // use APC if present
     if (!isset($argc) && function_exists('apc_fetch')) return apc_fetch($name);
@@ -221,8 +216,8 @@ class SRA_Cache {
 	 * @access public
 	 * @return boolean
 	 */
-  function setCache($name, &$val, $ttl=NULL, $maxAttempts=5) {
-    global $argc, $memcached, $memcachedKeys;
+  function setCache($name, &$val, $ttl=NULL, $maxAttempts=3) {
+    global $argc, $memcached;
     if (SRA_CACHE_DEBUG) SRA_Error::logError('SRA_Cache::setCache - invoked for "' . $name . '" with value "' . $val . '" ' . ($ttl ? 'and ttl "' . $ttl . '"' : ' and no ttl'), __FILE__, __LINE__);
     
     // use memcached if global variable $memcached exists
@@ -230,10 +225,11 @@ class SRA_Cache {
       // cache does not always set on the first try
       $cacheSet = FALSE;
       for($i=0; $i<$maxAttempts; $i++) {
-        if ($memcached->set($name, $val, $ttl)) {
+        $memcached->set($name, $val, $ttl);
+        $memcached->get($name);
+        if ($memcached->getResultCode() != Memcached::RES_NOTFOUND) {
           $cacheSet = TRUE;
-          if (is_array($memcachedKeys)) $memcachedKeys[] = $name;
-          break;          
+          break;
         }
         sleep(1);
       }
